Reposition the application title in the navbar
There’s a -15px margin-left set by Bootstrap on the application title 
that is pushing the crown a bit too far too the left.

Resetting the margin-left isn’t so easy without building up a complex
selector, so I’ve offset the left padding by 15px to account for this.