If a best bet is renamed, delete the old version
The notification has the query and type of the old version added, since
by the time the notification is received these are no longer stored in
the database.  If the old best bet no longer has any entries, a delete
is sent for it.  Otherwise, the old best bet is updated.