Fix N+1 query on Query list
Currently Rails is doing two SELECTs for each item in the query-list on
the landing page of this app. And there are a lot of items.

This commit refactors adds `best_bests` and `worst_bests` associations
for Query. This allows an `includes` to prevent the N+1 query.

Before: ActiveRecord: 365.1ms
After: ActiveRecord: 10.7ms