Show best bet links in order
Currently, the list of best bets is shown in DB order, not in order of
position. This happens both on the main list of best bets and in the
list for a particular query.

We have to use `sort_by` because the bets could be eager-loaded with
the query. Using an `order()` clause would make ActiveRecord perform an
N+1 query.

https://trello.com/c/Tk9QCE1i