Show iframe with search results on query page
To make it easier to see the effects of best/worst bets, we add an
iframe with the actual search results on the query page. The iframe
embeds the `gov.uk/search` page.

After a best bet is deleted and added it can take a little time (a
second or so) before elasticsearch returns the correct results. The
iframe is therefore not guaranteed to be current after an update. This
is mitigated by adding a callout with a warning and a button to refresh
the iframe.