Remove indirection in notifying rummager
When a bet/query is updated we need to send the changes to rummager.
Currently this is done via a MessageBus that calls listeners with the
attributes of the query.

This commit removes the MessageBus system in favour of calling the
RummagerNotifier directly. This reduces the complexity of the code
significantly without losing functionality.