Remove configurable search-index
search-admin currently allows users to choose between the `mainstream` and `government` search indexes for saving recommended links. Based on comments on https://github.com/alphagov/search-admin/pull/48 and a conversation with Tara Stockford, we’ve decided to send all recommended links to the `mainstream` index by default. This commit removes the option for the user to choose the index.