Add URL validation when adding recommended links
When creating a recommended link, search-admin currently accepts any string as a valid URL without validation. This means that invalid or relative URLs can be added.

This commit adds model-level validation to check that a submitted URL is a valid absolute URL with a scheme of either HTTP or HTTPS. Invalid URLs are rejected with an error message.