URL encode ElasticSearch bet ID
When sending a notification to rummager to delete a best/worst bet, search-admin does not correctly encode the bet ID in the URL. Therefore, when the ID includes spaces, the resulting URL fails a validation test.

This commit URL-encodes the ID when it is used to send a delete notification to rummager.