Replace Airbrake monitoring with Sentry
Use the govuk_app_config gem to report errors to Sentry instead of
Airbrake.

We're moving away from Errbit because it's an unsupported fork, isn't
meeting our needs very well and complicates the move to AWS. We're going
for Sentry because most people seem to like it, used by other teams in
GDS, it's modern and up-to-date and we can use a SaaS version.

https://trello.com/c/DJ7h8C1b/306-make-sure-the-apps-we-support-are-migrated-to-sentry
https://trello.com/c/INqrlNFs/61-update-search-admin