Refactor delete code and update encode to work with multiple words
Update the code so that saving a query with no bets will always try and delete it, this will happen even on creation. This overhead will help ensure the system stays insync with rummager.

Correct the delete url encoding. This was using `CGI.escape` and has been changed to `URI.escape` as it was not correctly escaping space characters, which resulted in 404 being returned, which the system took for a success.