Add travel advice bets importer
Add an importer which can handle rows of data
containing a search term and a help page link.
The importer creates bets bases on the link
if they don't already exist. It also adds a
best bet for the travel advice page associated
with the same country, and places this higher
than the help page bet.