Fix table filtering
Because additional columns have been added to the index page
table, the bets for each query needed to be their own row.
This led to a confusing layout where by the query field was repeated
for each bet. @kr8n3r had to rewrite the js to do two things:
 - hide the query for a bet, if the bet above it is for the same query.
 - unhide the query when the table is being filtered.