Delete 3-month-old expired or disabled bets
We might want to bring a bet back after it expires, but if it's been
left inactive for so long the chances of that are pretty slim.

Deleting the bets without deleting them from search-api is fine, as
they'll already have been removed from search-api by this point.

At the same time, delete empty queries too.  This removes cruft in the
search-admin UI and in the metasearch index.