Allow making a bet for a gov.uk link which has a recommended link
Normally we don't allow bets for `https;?/www.gov.uk/...` links, and
require the `/...` form.  But for paths on GOV.UK which don't have a
content item, like `https://www.gov.uk/alerts` or
`https://www.gov.uk/design-system`, we need to be able to add the full
URL.

Another approach would be to add dummy content items for these, so
every path on GOV.UK corresponds to some content.  But then we'd need
to add to the content item the information we want to appear in
search.