Revert require notify fix
According to the changelog: included in this release is a 'Revert Fix Zeitwerk compatibility issue'
which was made in this [commit](https://github.com/dxw/mail-notify/commit/59ca8bd7ba041230de3bb7cf9152e43cabd7da95).

We had fixed the issue that the Zeitwerk compatibility issue change introduced
[here](https://github.com/alphagov/search-admin/pull/828/commits/a69bf1a167822dc538f45649a9413d81a8e36273).

So, now we need to revert that change too.