Hold back a bad version of the `mail` rubygem.
Version 2.8.0 of the `mail` gem has the wrong filesystem permissions on
a few of its .rb files (seems to be the generated tables?), which breaks
the app when running as an unprivileged user. We therefore need to hold
back this version of the gem until a new version is released that fixes
the issue.

The upstream bug is https://www.github.com/mikel/mail/issues/1489. We
can remove this workaround once the bug is fixed.