Ensure Cucumber tests are always logged in
The best bets Cucumber tests also implicitly logged in a user for all
following tests instead of being isolated. This means now that they're
deleted, the other tests fail because there is no logged in user.

This adds a `Before` step to the Cucumber environment that logs in a
user before each test.