Use i18n for recommended links management
This makes it easier and cleaner to deal with diverging UI and code
naming ("recommended links" vs "external links") both for this feature
and for future features, and enables all content to be in one spot.

Also adds helpers to simplify some i18n behaviour that would normally
come "for free" when using Rails's built in form builder, but we need to
do explicitly due to using the GOV.UK Publishing Components directly.