Controls: Add ability to list and view controls
- Add `ControlsController` with `#index` and `#show` actions, plus view
  templates, routes, and translations
- Update the model translation helpers to take an optional argument
  specifying the model to use (but still maintaining the existing
  behaviour of using one inferred by the current controller)
- Add link to controls list page to application menu
- Add `context` and `context_inside` arguments to the shared
  `page_title` partial so we can use it to make clear what kind of
  action a control has on the show page

This also defines `#to_partial_path` on `Control::Actionable` to make it
easier to render partials for different actions. [`#to_partial_path`] is
how Rails's view layer figures out under the hood how to render an
ActiveModel instance passed to `render`.

By default, this returns a fully qualified partial path, so in this
case, Rails would want to render a FilterAction as
`app/views/control/filter_action/_filter_action.html.erb` (note the
singular control). That's a bit unwieldy, so this changes it to use just
the very last part of the class name, and when we use `render`, it'll
use `app/views/controls/_filter_action.html.erb` instead.

[`#to_partial_path`]: https://edgeapi.rubyonrails.org/classes/ActiveModel/Conversion.html#method-i-to_partial_path