Controls: Add `RemoteSynchronizable` concern
This adds a `RemoteSynchronizable` model concern used by `Control` (and
other models in the future) to persist themselves to DiscoveryEngine via
the `DiscoveryEngine::Control` client.

This concern uses ActiveRecord lifecycle callbacks to create/update/
delete a corresponding Control resource in Discovery Engine whenever its
local version changes.

Normally we would avoid using ActiveRecord callbacks to make network
calls to remote APIs, but as the core purpose of Search Admin is to
provide an interface to manage these resources on their various remote
APIs, this is part of its core domain.