Controls: Add `DiscoveryEngine::ControlClient`
This adds a client class for the Discovery Engine API, able to
synchronise a local `Control` instance.

It also adds a custom error class for this client and other future
clients, and some basic logic to try and figure out if a particular
error is caused by Discovery Engine disliking a filter expression, in
which case we can attach the error to the correct field.