Serving configs: Add initial `ServingConfig` model
This adds a model for serving configs. A serving config(uration) is an
endpoint on a Discovery Engine engine that serves results to clients
(such as Search API v2).

We will have a "default" serving config, used by Search API v2 to serve
live search results on GOV.UK. We will also be able to have one or more
secondary serving configs that we can use to try out configuration
changes (such as new controls).