Serving configs: Add Rake task to set up initial data
As of Feb 2025, the Discovery Engine Ruby client does not yet support
creating or deleting serving configs (due to those API endpoints only
being on the alpha API, which the Ruby client does not include).

Until this is supported, and we add the ability to create/delete serving
configs to Search Admin, we will operate with a fixed set of serving
configs as created by GOV.UK infrastructure Terraform, and imported by a
new Rake task.