Introduce `Current` per-request attributes singleton
This creates an idiomatic Rails per-request attributes singleton
`Current`, using [`ActiveSupport::CurrentAttributes`], and stores the
current user in it. This will allow us to access the currently active
user in cross-cutting concerns, without having to deeply pass it around.

`CurrentAttributes` is a "sharp tool" that should be used with caution
as it creates global state for a request, but is an ideal tool for
dealing with auditing and logging concerns (which we'll soon introduce)
where you need to have global awareness of certain request attributes
outside of just the controller layer.

We have prior art for this pattern, for example in Whitehall.

[`ActiveSupport::CurrentAttributes`]: https://edgeapi.rubyonrails.org/classes/ActiveSupport/CurrentAttributes.html