Add `Engine` and `DataStore` models
These are plain Ruby objects representing the concepts of engine and
data store on Discovery Engine.

Our architecture only has a single one of each, and that is unlikely to
change in the medium term, so these are not persisted in the database
but rather just have a single default instance available as `.default`.