Add initial `ServingConfig` model
This represents a serving config(uration) on Discovery Engine, which is
an endpoint on an engine that can be used for querying. The main purpose
of serving configs is to have controls attached which modify their
behaviour, and each of an engine's serving configs can have zero-to-many
of that engine's controls attached. Having several serving configs means
we can test out different sets of controls and how they affect search
results.

One serving config (`govuk_default`) is special in that it is the
designated default. This will mirror the hardcoded default serving
configuration in Search API v2 (which is currently the legacy serving
config `default_search` auto-generated by Discovery Engine).

Serving configs have a display name and remote resource identifier (like
other Discovery Engine resources), and a boolean flag that determines
whether users of Search Admin should be able to attach controls to it
through the UI (as opposed to just allowing programmatic modification by
application code).

As we cannot currently create or delete serving configs through the
Discovery Engine Ruby client, initial creation of three serving configs
has been done in `govuk-infrastructure` [1] and we use a Rake task to
create the model records on this end. Depending on whether these three
are enough for the long term, we may or may not want to add the ability
to create, edit, and delete serving configs to Search Admin in the
future.

[1]: https://github.com/alphagov/govuk-infrastructure/pull/1680