Add initial `ServingConfig` model
This represents a serving config(uration) on Discovery Engine, which is
an endpoint on an engine that can be used for querying. The main purpose
of serving configs is to have controls attached which modify their
behaviour, and each of an engine's serving configs can have zero-to-many
of that engine's controls attached. Having several serving configs means
we can test out different sets of controls and how they affect search
results.

Serving configs have one of three use cases:
- `live` serving configs are used somewhere in our search stack to
  provide results to the public
- `preview` serving configs are for use by Search Admin users to test
  new configurations (for example, when adding new controls)
- `system` serving configs are for internal programmatic use and cannot
  be modified through the Search Admin UI (for example, we have an
  "empty" serving config to test having no controls at all)

Serving configs also have a display name and remote resource identifier
(like other Discovery Engine resources).

As we cannot currently create or delete serving configs through the
Discovery Engine Ruby client, initial creation of three serving configs
has been done in `govuk-infrastructure` [1] and we use a Rake task to
create the model records on this end. Depending on whether these three
are enough for the long term, we may or may not want to add the ability
to create, edit, and delete serving configs to Search Admin in the
future.

[1]: https://github.com/alphagov/govuk-infrastructure/pull/1680