Add comments to controls
This adds a `comment` field to controls, analogous to recommended links,
allowing users to explain the purpose of a control.

As MySQL (currently as of v8) doesn't support default values for `TEXT`
columns, we have to do a bit of a dance to add a new non-`NULL` column
of that type to a table.