Move recommended link API presentation into model
This brings the generation of a Publishing API content item for a
recommended link from a separate presenter into the `RecommendedLink`
model itself. As the core purpose of Search Admin is to be an app
managing state on external systems, we consider this representation to
be part of the core domain model and not a separate pure presentation
concern.