Update RemoteSynchronizable documentation
This concern was reverted in d045d1d to the previous implementation where the
client is explicitly set as a class variable rather than by a method
call, but it looks like the documentation still reflects the previous
way.

Similarly, it no longer uses lifecycle callbacks but relies on these
methods to be explicitly called.