Simplify API client error handling
The way we handle errors in API clients is a bit too complex, and too
coupled to ActiveRecord (it sets model errors, possibly based on trying
to parse an error string, and raises its own `ClientError`).

This changes the clients to not rescue any errors from the remote API,
but instead letting them bubble up to the caller, and moves the setting
of record errors to the ActiveRecord-specific `RemoteSynchronizable`
concern, paving the way for clients to be able to work with POROs as
well.