Add worker to import completion denylist entries
This adds a worker to import the local state of completion denylist
entries to Discovery Engine.

It uses the (commonly used across GOV.UK) `with_advisory_lock` gem to
ensure only a single worker runs at a time, in the unlikely scenario
several changes are made simultaneously or in quick succession.